add cart_api

+++ app/api/utils/function.go
@@ -2,10 +2,10 @@ package utils
 
 import "context"
 
-func GetUserIdFromCtx(ctx context.Context) int32 {
+func GetUserIdFromCtx(ctx context.Context) uint32 {
 	userId := ctx.Value(SessionUserId)
 	if userId == nil {
 		return 0
 	}
-	return userId.(int32)
+	return uint32(userId.(float64))
 }
